grub2: Add CVE whitelist entries for issues fixed in 2.06
We're using a pre-release version of 2.06 so these issues are fixed but
continue to show up in the checks since it is pre-2.06 and the CPE
entries are "before but excluding 2.06".

Adding these will clean up CVE reports until the 2.06 release comes out.

(From OE-Core rev: 2467ab1554bee3a431636046735e8e369e865bc6)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>